Create Module 12 Self Prompt
In my Intro to Python course, we were assigned to design our own code that:
- reads or writes to a file
-utilizes a list or dictionary
- involves user input
- involves functions
- involves ifs or loops
For my project, I coded a txt file of a teacher's student's grades. The code takes the txt file and converts the numbers to a dictionary by grade number. It tells how many students got each grade. It also calculates the average.